+++ concordium-sdk/src/main/java/com/concordium/sdk/Client.java
@@ -16,7 +16,6 @@ import com.concordium.sdk.responses.cryptographicparameters.CryptographicParamet
 import com.concordium.sdk.responses.transactionstatus.TransactionStatus;
 import com.concordium.sdk.transactions.*;
 import com.google.protobuf.ByteString;
-import com.vdurmont.semver4j.Semver;
 import concordium.ConcordiumP2PRpc;
 import concordium.P2PGrpc;
 import io.grpc.ManagedChannel;
@@ -257,12 +256,19 @@ public final class Client {
     }
 
     /**
-     * Gets the Semantic Version of the Peer Software / Node
-     * @return {@link Semver} of the Peer / Node
+     * Gets the total number of packets sent.
+     * @return Total number of packets sent.
      */
-    public Semver getVersion() {
-        val value = server().peerVersion(ConcordiumP2PRpc.Empty.newBuilder().build()).getValue();
-        return new Semver(value);
+    public long getTotalSent() {
+        return server().peerTotalSent(ConcordiumP2PRpc.Empty.newBuilder().build()).getValue();
+    }
+
+    /**
+     * Gets the Semantic Version of the Peer Software / Node
+     * @return Version of the Peer / Node
+     */
+    public String getVersion() {
+        return server().peerVersion(ConcordiumP2PRpc.Empty.newBuilder().build()).getValue();
     }
 
     /**