+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed bug in `UpdateEnqueuedResult` which parsed `effectiveTime` in wrong.
 - Added support for `Uptime` query.
+- Added support for query `PeerTotalSent`
 - Added support for query `PeerVersion`
 
 ## 3.0.0